==> SVG_LogoGenerator/animate_logos_main_adapted/notebooks/1_svg_preprocessing.ipynb <==
### Preprocessing

import os
os.chdir("..")

import pandas as pd
import numpy as np
import pickle

from src.data import insert_ids_in_folder, rename_logos, expand_viewbox_in_folder, get_svg_meta_data 
from src.preprocessing import decompose_logos_in_folder, get_path_relevance, apply_embedding_model_to_svgs, encode_svg, decode_z
from src.features import get_style_attributes_folder
from src.animations import create_animated_svg, create_random_animations

#### 1 Insert IDs and Expand Viewbox
An ID called animation_id is inserted for every path in a SVG and the viewbox is expanded.

insert_ids_in_folder("data/rawdata_logo_dataset", "data/svgs_preprocessed")

expand_viewbox_in_folder('data/svgs_preprocessed', 50, "data/svgs_preprocessed_expanded")

#### 2 Decompose SVGs
Decompose SVGs by saving every path in a seperate SVG file. Decomposed SVGs are save to data/decomposed_svgs.

decompose_logos_in_folder("data/svgs_preprocessed_expanded")

#### 4 Get SVGs Metadata

pd.set_option('display.max_rows', None)

meta_data = get_svg_meta_data("data/svgs_preprocessed_expanded")

meta_data.to_csv("data/meta_data/svgs_meta.csv", index=False)

#### 5 Get Path Embeddings
Apply hierarchical generative network to produce latent vector z

apply_embedding_model_to_svgs(data_folder="data/decomposed_svgs", split_paths=True, save=True)
