==> SVG_LogoGenerator/animate_logos_main_adapted/notebooks/3a_surrogate_model_create_data.ipynb <==
## Create random animations

NOTE: This notebook shows how to create random animations. These random animations were labelled afterwards and the resulting data saved to data/surrogate_model/animation_label.pkl. Since running this notebook again would lead to new random animations with unknown labels, you can skip this notebook and proceed with notebook 3a_surrogate_model_transform_data.ipynb.

The random animated SVGs are saved to the folder data/animated_svgs.

import os
os.chdir("..")

import pickle
import warnings
import pandas as pd
from src.animations.create_random_animations import create_random_animations, combine_dataframes, create_backend_mapping_df
warnings.filterwarnings('ignore')
pd.set_option('display.max_rows', 10)

### 1. Create Animations
Animations are not completely random. Probabilies of path labeling are considered.

##### Specify number of animations per logo

nb_animations = 5

##### Animation vectors are saved as dataframe in animate_logos/data/animated_svgs_dataframes
If split_df=true, animation vectors are saved to multiple dataframes (one dataframe per logo).  
If split_df=false, animation vectors are saved to one dataframe and returned.

If nb_animations very high, use split_df=true.

split_df = False

create_random_animations("data/svgs_preprocessed_expanded", nb_animations, split_df)

### 2. Concatenat dataframes
This function concatenates all pkl files in a given folder. Please move or delete all pkl files that you don't want to concantenate.

df_combined = combine_dataframes("data/animated_svgs_dataframes")
df_combined.head(20)

### 3. Create backend mapping

backend_mapping_df = create_backend_mapping_df(df_combined)
backend_mapping_df

# save backend mapping
backend_mapping_df.to_csv('data/animated_svgs_dataframes/backend_mapping.csv', index=False)

### 4. Insert label in concatenated dataframe
The following function can be used to merge the corresponding label to the random animations.

def merge_label(df_combined, label_path, save=False):
    label = pd.read_csv(label_path, skiprows=[1], dtype={"alias": object, "animation_file": object, "animation_id": int,
                                                       "logo_id": int, "path_0_rating": object, "path_1_rating": object,
                                                       "path_2_rating": object, "path_3_rating": object, 
                                                       "path_4_rating": object, "path_5_rating": object, 
                                                       "path_6_rating": object, "path_7_rating": object,
                                                       "time": object})
    
    label_melt = pd.melt(label, id_vars=["alias","animation_file", "animation_id", "logo_id", "time"],
                                  value_vars=["path_0_rating", "path_1_rating", "path_2_rating", "path_3_rating", "path_4_rating", "path_5_rating", "path_6_rating", "path_7_rating"],
                                  var_name="path", value_name="label").assign(order_id=label_melt.path.str.extract("(\d+)")).astype({"order_id": 'int'})
    
    label_melt_id = label_melt.assign(order_id=label_melt.path.str.extract("(\d+)")).astype({"order_id": 'int'}).rename(columns={"animation_id":"animation_number"}).drop("path",1)
    
    df_combined_id = df_combined.assign(logo_id=df_combined.file.str.extract("logo_(\d+)"),
                                        animation_number=df_combined.file.str.extract("animation_(\d+)")).astype({"logo_id": 'int',"animation_number": 'int'})
     
    # Combine dataframes and merge data
    label_path_merged = pd.merge(df_combined_id, label_melt_id, how='left', on=["logo_id","animation_number","order_id"])
    
    if save:
        # save dataframe
        output = open(f"data/surrogate_model/animation_label.pkl", 'wb')
        pickle.dump(label_path_merged, output)
        output.close()
    
    return label_path_merged
