==> SVG_LogoGenerator/animate_logos_main_adapted/notebooks/5b_generation_model_entmoot_evaluation.ipynb <==
import os, pickle, random
import pandas as pd
import numpy as np
from datetime import datetime
from collections import Counter

os.chdir('..')
from src.models.genetic_algorithm import *
from src.models.train_animation_predictor import *
from src.models import config
from src.models.ordinal_classifier_fnn import *
from src.models.genetic_algorithm import create_animation_vector
from src.animations.insert_animation import create_animated_svg
from src.models.entmoot_functions import *

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

## 1. Load test data for evaluation

test_data = pd.read_csv('data/path_selector/path_selector_test.csv')

test_data.head()

test_data = retrieve_m1_predictions(test_data)

test_data = retrieve_animation_midpoints(test_data, drop=True)

scaler = pickle.load(open(config.scaler_path, 'rb'))
test_data[config.sm_features] = scaler.transform(test_data[config.sm_features])

test_data.head()

test_data.to_csv('data/entmoot_optimization/evaluation_data.csv', index=False)

len(test_data)

## 2. Apply Entmoot optimizer to test data

# Set seeds in order to reproduce results
torch.manual_seed(73)
random.seed(73)
np.random.seed(73)

# Load test data for checking evaluation procedure
test_data = pd.read_csv('data/entmoot_optimization/evaluation_data.csv')
path_vectors = test_data[config.sm_features].values.tolist()

with open("models/entmoot_optimizer_final.pkl", "rb") as f:
    optimizer = pickle.load(f)

# Load surrogate model for function evaluations
func = SurrogateModelFNN()

# Predict animation vectors
an_vec_preds = []
score_preds = []
for i in range(len(path_vectors)):
    opt_x, opt_y, _ = entmoot_predict(optimizer, func, path_vectors[i], n_calls=1)
    an_vec_preds.append(opt_x)
    score_preds.append(-opt_y)
    if (i+1) % 10 == 0:
        print(f'Finished animation prediction for path {i+1}.')

test_data['animation_vector'] = an_vec_preds

test_data['reward'] = score_preds

test_data.head()

test_data['reward'].mean()

test_data['reward'].value_counts()
