==> SVG_LogoGenerator/animate_logos_main_adapted/notebooks/6_logo_animation_pipeline.ipynb <==
## SVG Pipeline

import os
# os.chdir("..")
from src.pipeline import * 

cd ..

%pip install entmoot

ls

%pip install --upgrade torch

#### Load Logo

logo = Logo(data_dir="data/uploads/Uni_Mannheim_Siegel.svg")

logo = Logo(data_dir="/Users/samailguliyev/Downloads/animate_logos-main-UNTOUCHED/data/svgs/easeus.svg")

#### Print logo information

logo.print_logo_information()

#### Create Animations
Animations are saved to the same folder.

# Genetic Algorithm
logo.animate(model='ga')

# Entmoot optimizer
logo.animate(model='opt')
